cel-worker/requirements.txt
flask
celery
flask-mysqldb
mysql-connector-python
sqlalchemy
redis